Hotfix/TRANSLATE-5317 fix language switch in grid

* TRANSLATE-5317 UI language switch in grid broken

* * implement new locale-switching

* -

* -

* -

* -

* * added doc

* Merge branch 'master' into hotfix/TRANSLATE-5317-fix-language-switch-in-grid

* Merge branch 'master' into hotfix/TRANSLATE-5317-fix-language-switch-in-grid

* * finish locale-switching, add new endpoint for user-locale switch, remove global "locale" param

* * add locale option

* Merge commit '1119e399bc22b449fd22d163e6feb0a332a42663' into hotfix/TRANSLATE-5317-fix-language-switch-in-grid


Approved-by: Thomas Lauria

+++ application/modules/default/Controllers/LoginController.php
@@ -374,7 +374,7 @@ class LoginController extends ZfExtended_Controllers_Login
 
         try {
             $user = $userRepository->get(
-                Auth::getInstance()->getUser()->getId()
+                (int) Auth::getInstance()->getUser()->getId()
             );
         } catch (InexistentUserException) {
             return;
